+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/fit/baseTypes/BaseTypeInt.java
@@ -42,7 +42,10 @@ public class BaseTypeInt implements BaseTypeInterface {
             invalidate(byteBuffer);
             return;
         }
-        long l = (long) ((((Number) o).longValue() + offset) * scale);
+        // Use doubleValue() rather than longValue() so that fractional inputs survive the
+        // multiply by scale instead of being truncated first. For Integer/Long inputs the
+        // result is identical.
+        long l = (long) ((((Number) o).doubleValue() + offset) * scale);
         if (l < min || l > max) {
             invalidate(byteBuffer);
             return;
